Распил по файлам виджетов sightCard
