+++ lib/ui/sight/sight_list/sight_list_page.dart
@@ -6,6 +6,13 @@ import 'package:places/domain/sight/sight.dart';
 import 'package:places/mocks.dart';
 import 'package:places/ui/icons/svg_icons.dart';
 import 'package:places/ui/sight/sight_card/sight_card.dart';
+import 'package:places/ui/sight/sight_card/widgets/actions.dart';
+import 'package:places/ui/sight/sight_card/widgets/body.dart';
+import 'package:places/ui/sight/sight_card/widgets/details_text.dart';
+import 'package:places/ui/sight/sight_card/widgets/header.dart';
+import 'package:places/ui/sight/sight_card/widgets/image.dart';
+import 'package:places/ui/sight/sight_card/widgets/name_text.dart';
+import 'package:places/ui/sight/sight_card/widgets/type_text.dart';
 
 class SightListPage extends StatefulWidget {
   const SightListPage({Key? key}) : super(key: key);
